@@ -10,41 +10,41 @@
     "npm": ">= 2.1"
   },
   "dependencies": {
-    "babel": "4.7.1",
+    "babel": "4.7.12",
     "bootstrap": "3.3.2",
     "eventemitter3": "0.1.6",
     "express": "4.12.2",
     "flux": "2.0.1",
     "front-matter": "1.0.0",
     "jade": "1.9.2",
-    "lodash": "3.4.0",
-    "react": "0.12.2",
-    "superagent": "0.21.0"
+    "lodash": "3.5.0",
+    "react": "0.13.0",
+    "superagent": "1.1.0"
   },
   "devDependencies": {
     "autoprefixer-loader": "^1.2.0",
-    "babel-core": "^4.7.1",
-    "babel-eslint": "^1.0.13",
+    "babel-core": "^4.7.12",
+    "babel-eslint": "^2.0.0",
     "babel-loader": "^4.1.0",
-    "browser-sync": "^2.2.2",
+    "browser-sync": "^2.2.4",
     "css-loader": "^0.9.1",
     "del": "^1.1.1",
-    "eslint": "^0.15.1",
-    "eslint-loader": "^0.4.0",
+    "eslint": "^0.16.0",
+    "eslint-loader": "^0.6.0",
     "gulp": "^3.8.11",
     "gulp-autoprefixer": "^2.1.0",
     "gulp-cache": "^0.2.8",
-    "gulp-changed": "^1.1.1",
+    "gulp-changed": "^1.2.1",
     "gulp-csscomb": "^3.0.3",
-    "gulp-gh-pages": "^0.4.0",
+    "gulp-gh-pages": "^0.5.0",
     "gulp-htmlmin": "^1.1.1",
     "gulp-if": "^1.2.5",
     "gulp-imagemin": "^2.2.1",
-    "gulp-jsbeautifier": "^0.0.5",
+    "gulp-jsbeautifier": "^0.0.6",
     "gulp-less": "^3.0.1",
     "gulp-load-plugins": "^0.8.1",
-    "gulp-minify-css": "^0.5.1",
-    "gulp-plumber": "^0.6.6",
+    "gulp-minify-css": "^1.0.0",
+    "gulp-plumber": "^1.0.0",
     "gulp-rename": "^1.2.0",
     "gulp-render": "^0.2.2",
     "gulp-replace": "^0.5.3",
@@ -54,14 +54,14 @@
     "jest-cli": "^0.4.0",
     "less": "^2.4.0",
     "less-loader": "^2.1.0",
-    "minimist": "^1.1.0",
+    "minimist": "^1.1.1",
     "protractor": "^1.8.0",
     "psi": "^1.0.6",
-    "react-tools": "^0.12.2",
+    "react-tools": "^0.13.0",
     "run-sequence": "^1.0.2",
     "style-loader": "^0.8.3",
     "url-loader": "^0.5.5",
-    "webpack": "^1.7.2",
+    "webpack": "^1.7.3",
     "webpack-dev-server": "^1.7.0"
   },
   "jest": {
