@@ -10,15 +10,17 @@
 
 import React from 'react';
 
-export default React.createClass({
-
-  propTypes: {
-    body: React.PropTypes.string.isRequired
-  },
+export default class HomePage extends React.Component {
 
   render() {
-    return <div className="ContentPage"
-      dangerouslySetInnerHTML={{__html: this.props.body}} />;
+    return (
+      <div className="ContentPage"
+        dangerouslySetInnerHTML={{__html: this.props.body}} />
+    );
   }
 
-});
+}
+
+HomePage.propTypes = {
+  body: React.PropTypes.string.isRequired
+};
