@@ -12,7 +12,7 @@
 
 import React from 'react';
 
-export default React.createClass({
+export default class NotFoundPage extends React.Component {
 
   render() {
     return (
@@ -23,4 +23,4 @@ export default React.createClass({
     );
   }
 
-});
+}
