@@ -16,7 +16,7 @@ import assign from 'react/lib/Object.assign';
  * A singleton that operates as the central hub for application updates.
  * For more information visit https://facebook.github.io/flux/
  */
-var Dispatcher = assign(new Flux.Dispatcher(), {
+let Dispatcher = assign(new Flux.Dispatcher(), {
 
   /**
    * @param {object} action The details of the action, including the action's
@@ -44,4 +44,4 @@ var Dispatcher = assign(new Flux.Dispatcher(), {
 
 });
 
-module.exports = Dispatcher;
+export default Dispatcher;
