@@ -10,11 +10,9 @@
 
 import keyMirror from 'react/lib/keyMirror';
 
-var PayloadSources = keyMirror({
+export default keyMirror({
 
   VIEW_ACTION: null,
   SERVER_ACTION: null
 
 });
-
-module.exports = PayloadSources;
